+++ test/index.js
@@ -8,16 +8,31 @@ const less = ruleTester(validator.rule, validator.ruleName, {
 const sass = ruleTester(validator.rule, validator.ruleName, {
     postcssOptions: { syntax: require('postcss-scss') }
 });
+const invalid = (prop, line, column) => {
+    const message = messages.invalid(prop);
+
+    if (typeof line !== 'number' && typeof column !== 'number') {
+        return message;
+    }
+
+    return { message, line, column };
+}
 
 // base test
 css(null, function(tr) {
     tr.ok('.foo { color: red }');
     tr.ok('.foo { color: #123456 }');
-    tr.notOk('.foo { color: red green }', messages.invalid('color'));
-    tr.notOk('.foo { color: 1 }', messages.invalid('color'));
-    tr.notOk('.foo { color: #12345 }', messages.invalid('color'));
+    tr.notOk('.foo { color: red green }', invalid('color'));
+    tr.notOk('.foo { color: 1 }', invalid('color'));
+    tr.notOk('.foo { color: #12345 }', invalid('color'));
     tr.notOk('.foo { color: &a }', messages.parseError('&a'));
     tr.notOk('.foo { baz: 123 }', 'Unknown property `baz`');
+});
+
+// loc
+css(null, function(tr) {
+    tr.notOk('.foo { color: red green }', invalid('color', 1, 19));
+    tr.notOk('.foo {\n  width: 10px;\n  color: red green;\n}', invalid('color', 3, 14));
 });
 
 // ignore values with less extenstions
@@ -81,7 +96,7 @@ css({ ignoreValue: "^patternToIgnore$", ignore: ['bar'] }, function(tr) {
     tr.ok('.foo { color: #fff }');
     tr.ok('.foo { color: patternToIgnore }');
     tr.ok('.foo { bar: notMatchingPattern }');
-    tr.notOk('.foo { color: notMatchingPattern }', messages.invalid('color'));
+    tr.notOk('.foo { color: notMatchingPattern }', invalid('color'));
     tr.notOk('.foo { foo: patternToIgnore }', 'Unknown property `foo`');
     tr.notOk('.foo { foo: notMatchingPattern }', 'Unknown property `foo`');
 });
@@ -106,10 +121,24 @@ css({
     tr.ok('.bar { bar: my-fn(10px) }');
     tr.notOk('.baz { baz: my-fn(10px) }', 'Unknown property `baz`');
     tr.ok('.qux { qux: 10px }');
-    tr.notOk('.foo { color: my-fn(10px) }', messages.invalid('color'));
+    tr.notOk('.foo { color: my-fn(10px) }', invalid('color'));
     tr.ok('.foo { color: darken(white, 5%) }');
     tr.ok('.foo { color: white }');
-    tr.notOk('.foo { color: darken(white, .05) }', messages.invalid('color'));
+    tr.notOk('.foo { color: darken(white, .05) }', invalid('color'));
     tr.ok('.foo { relax: white }');
     tr.ok('.foo { relax: 10px solid whatever }');
+});
+
+css({
+    properties: {
+        composes: '<custom-ident>+ [ from [ <string> | global ] ]?'
+    }
+}, function(tr) {
+    tr.ok('.foo { composes: className }');
+    tr.ok('.foo { composes: classNameA classNameB }');
+    tr.ok('.foo { composes: classNameA from "foo.css" }');
+    tr.ok('.foo { composes: classNameA from global }');
+    tr.ok('.foo { composes: classNameA classNameB from "foo.css" }');
+    tr.notOk('.foo { composes: from "foo.css" }', invalid('composes', 1, 23));
+    tr.notOk('.foo { composes: classNameA "foo.css" }', invalid('composes', 1, 29));
 });